==> scripts/7_analyses.Rmd <==
---
title: "Prior collaboration networks"
author: "Bas Hofstra"
date: '2023-20-04'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Staging the script

What we want is a collaboration network that tells us whether the crew in current movies worked together before!

```{r staging, eval = TRUE}
#staging the script
rm(list = ls())

# custom function to check for packages / installs
fpackage.check <- function(packages) {
  lapply(packages, FUN = function(x) {
    if (!require(x, character.only = TRUE)) {
      install.packages(x, dependencies = TRUE)
      library(x, character.only = TRUE)
    }
  })
}


# declare packages
packages = c("tidyverse", "statnet", "igraph", "reshape2", "readr", "dplyr", "stringr", "car", "dotwhisker", "cowplot")


# load/ install pacakges
fpackage.check(packages)

```



```{r loading, eval = TRUE}

setwd('..') # somehow WD was one too high?
# read the data
df <- read.csv("data/final_results/ChangingThePicture_mastersheet.csv", sep = ";")
collabs <- read.csv("data/final_results/collabnets.csv", sep = ",")
#collabs2 <- read.csv("data/final_results/collabnets2.csv", sep = ",")



df <- left_join(df, collabs[, c("tconst", "collabdens", "collabcentr")], by = c("tconst" = "tconst"))

# tconst == movies
# nconst == actor

```

## Staging some identifiers we need later on

So we need the number of movies, the number of years, and a vector with movie identifiers.

```{r varstaging, eval = TRUE}


df$Metascore[df$Metascore == "N/A"] <- NA

df$BT.score[df$df$BT.score == "N/A"] <- NA
df$BT.id[df$df$BT.id == "N/A"] <- NA
df$BT.dubious[df$BT.dubious == "N/A"] <- NA
df$Awards[df$BT.dubious == "N/A"] <- NA
df$Metascore[df$Metascore == ""] <- NA


# number of wins and nominations in awards
df$winsnoms <- sapply(str_extract_all(df$Awards, "[[:digit:]]+"),
  function(x) ifelse(identical(x, character(0)),NA,sum(as.numeric(x))))


df <- df[, c("tconst", "Metascore", "imdbRating", "BoxOffice", "winsnoms", 
             "releaseYear", 
             "meanPreviousIMDbRatingDirector", 
             "crewCount", "collabdens", "collabcentr", 
             "nr_nodes", "density", "centralization_degree",
             "dir_percentage_female", "percentageFemaleCrew", "char_percentage_female", 
             "BT.score", "BT.dubious"
             )]

names(df) <- c("tconst", "metascore", "imdbrating", "boxoffice", "winsnoms", 
             "year", 
             "directorrating", 
             "ncrew", "collabdens", "collabcentr", 
             "nchars", "chardens", "charcentr", 
             "pfemale_dir", "pfemale_crew", "pfemale_chars", 
             "bechdelscore", "bechdeldubious"
             )

for (i in 2:length(df)) {
  
  df[[i]] <- as.numeric(df[[i]])
  
}

df$female_dir <- ifelse(df$pfemale_dir != 0, 1, 0)
df$fullcrew <- ifelse(df$ncrew == 10, 1, 0)
df$bf2000 <- ifelse(df$year < 2000, 0, 1)
df$bechdelscore <- ifelse(df$bechdelscore != 3, 0, 1)
```


```{r depdistr, eval = TRUE}
df$metascore <- df$metascore/10
meta <- ggplot(df, aes(x=metascore)) + geom_density() + theme_minimal() + xlab("Metascore") + ylab("Density")
imdb <- ggplot(df, aes(x=imdbrating)) + geom_density() + theme_minimal() + xlab("IMDb rating") + ylab("Density")
ggplot(df, aes(x=boxoffice)) + geom_density() + theme_minimal()
ggplot(df, aes(x=winsnoms)) + geom_density() + theme_minimal()


df$lwinsnoms <- log10(df$winsnoms)

box <- ggplot(df, aes(x=lboxoffice)) + geom_density() + theme_minimal() + xlab("log(Box office $)") + ylab("Density")
wins <- ggplot(df, aes(x=lwinsnoms)) + geom_density() + theme_minimal() + xlab("log(Wins and nominations)") + ylab("Density")

plot_grid(meta, imdb, box, wins, nrow = 2, labels = c("A","B", "C", "D"))

```


```{r depstaging, eval = TRUE}



m1a <- lm(metascore ~ bf2000 +
             ncrew + collabdens + collabcentr +
             pfemale_crew 
             , data = df)

m1b <- lm(metascore ~ bf2000 +
             ncrew + collabdens + collabcentr +
             nchars + chardens + charcentr +
             pfemale_crew + pfemale_chars + bechdelscore 
             , data = df)

m2a <- lm(imdbrating ~ bf2000 +
             ncrew + collabdens + collabcentr +
             pfemale_crew 
             , data = df)

m2b <- lm(imdbrating ~ bf2000 +
             ncrew + collabdens + collabcentr +
             nchars + chardens + charcentr +
             pfemale_crew + pfemale_chars + bechdelscore 
             , data = df)

m3a <- lm(lboxoffice ~ bf2000 +
             ncrew + collabdens + collabcentr +
             pfemale_crew 
             , data = df)

m3b <- lm(lboxoffice ~ bf2000 +
             ncrew + collabdens + collabcentr +
             nchars + chardens + charcentr +
             pfemale_crew + pfemale_chars + bechdelscore
             , data = df)

m4a <- lm(lwinsnoms ~ bf2000 +
             ncrew + collabdens + collabcentr +
             pfemale_crew 
             , data = df)

m4b <- lm(lwinsnoms ~ bf2000 +
             ncrew + collabdens + collabcentr +
             nchars + chardens + charcentr +
             pfemale_crew + pfemale_chars + bechdelscore
             , data = df)

summary(m1a)
summary(m1b)
summary(m2a)
summary(m2b)
summary(m3a)
summary(m3b)
summary(m4a)
summary(m4b)


f1 <- dwplot(list(m1a, m1b), ci = 0.95) %>%
        relabel_predictors(c(
                bf2000 = "After 2000",
                ncrew = "# crew",
                collabdens = "Collaboration density",
                collabcentr = "Collaboration centralization",
                pfemale_crew = "% women crew",
                chardens = "Character density",
                charcentr = "Character centralization",
                nchars = "# characters",
                pfemale_chars = "% women chararacters",
                bechdelscore = "Bechdel score")) + 
  theme_minimal() +
    scale_color_hue(breaks = c("Model 2", "Model 1"), labels = c('Production + Narrative', 'Production')) +
  theme(legend.title = element_blank(),
        legend.position = "top",
        legend.text = element_text(size=5),
        axis.text.y = element_text(size=5),
        plot.title = element_text(size = 8)) + ggtitle("Metascore")
       


f2 <- dwplot(list(m2a, m2b), ci = 0.95) %>%
        relabel_predictors(c(
                bf2000 = "a",
                ncrew = "b",
                collabdens = "c",
                collabcentr = "d",
                pfemale_crew = "e",
                chardens = "f",
                charcentr = "g",
                nchars = "h",
                pfemale_chars = "i",
                bechdelscore = "j")) + 
  theme_minimal() +
    scale_color_hue(breaks = c("Model 2", "Model 1"), labels = c('Production + Narrative', 'Production')) +
  theme(legend.title = element_blank(),
        legend.text = element_text(size=5),
        legend.position = "top",
        axis.text.y = element_text(color="white", size = 5),
        plot.title = element_text(size = 8)) + ggtitle("IMDb rating") 
       
f3 <- dwplot(list(m3a, m3b), ci = 0.95) %>%
        relabel_predictors(c(
                bf2000 = "After 2000",
                ncrew = "# crew",
                collabdens = "Collaboration density",
                collabcentr = "Collaboration centralization",
                pfemale_crew = "% women crew",
                chardens = "Character density",
                charcentr = "Character centralization",
                nchars = "# characters",
                pfemale_chars = "% women chararacters",
                bechdelscore = "Bechdel score")) + 
  theme_minimal() +
    scale_color_hue(breaks = c("Model 2", "Model 1"), labels = c('Production + Narrative', 'Production')) +
  theme(legend.title = element_blank(),
        legend.text = element_text(size=5),
        axis.text.y = element_text(size=5),
        legend.position = "top",
        plot.title = element_text(size = 8)) + ggtitle("log(Box office $)")
       
f4 <- dwplot(list(m4a, m4b), ci = 0.95) %>%
        relabel_predictors(c(
                bf2000 = "a",
                ncrew = "b",
                collabdens = "c",
                collabcentr = "d",
                pfemale_crew = "e",
                chardens = "f",
                charcentr = "g",
                nchars = "h",
                pfemale_chars = "i",
                bechdelscore = "j")) +
  theme_minimal() +
    scale_color_hue(breaks = c("Model 2", "Model 1"), labels = c('Production + Narrative', 'Production')) +
  theme(legend.title = element_blank(),
        legend.text = element_text(size=5),
        legend.position = "top",
        axis.text.y = element_text(color="white", size = 5),
        plot.title = element_text(size = 8))+ ggtitle("log(Wins and nominations)")
fig2 <- plot_grid(f1, f2, f3, f4, nrow = 2, labels = c("A", "B", "C", "D"), rel_widths = c(0.55, 0.45))

# save
ggsave("results_fig2.pdf", plot = fig2, device = "pdf",
       scale = 1, width = 7, height = 5, units = c("in"),
       dpi = "retina")


dwplot(m2b, ci = 0.95) + theme_minimal()
dwplot(m3a, ci = 0.95) + theme_minimal()
dwplot(m3b, ci = 0.95) + theme_minimal()
dwplot(m4a, ci = 0.95) + theme_minimal()
dwplot(m4b, ci = 0.95) + theme_minimal()

vif(m1a)
vif(m1b)
vif(m2a)
vif(m2b)
vif(m3a)
vif(m3b)
vif(m4a)
vif(m4b)

plot(m1a$residuals)
plot(m1b$residuals)
plot(m2a$residuals)
plot(m2b$residuals)
plot(m3a$residuals)
plot(m3b$residuals)
plot(m4a$residuals)
plot(m4b$residuals)


```
